Add the skip code

+++ preprocessing/run_preprocessor.py
@@ -24,13 +24,22 @@ PREPROCESSED_DIRECTORY = Property.get_property("preprocessed_directory")
 #     cv2.imwrite(combined_image_path, combined_frame)
 
 def run(m_data):
+    # Set the target image path
     filename_with_extension, label, path = m_data
-    # video_frames = Preprocessor.read_video(f"{path}/{filename_with_extension}")
-    video_frames = Preprocessor.read_video_and_extract_face(f"{path}/{filename_with_extension}")
-    combined_frame = Preprocessor.combine_video_frames(video_frames)
-
     filename, extension = os.path.splitext(filename_with_extension)
     combined_image_path = f"{PREPROCESSED_DIRECTORY}/{label}/{filename}.jpg"
+    if os.path.exists(combined_image_path):
+        print(f"\t* Skipped {combined_image_path}")
+        return
+
+    # Get frames
+    # video_frames = Preprocessor.read_video(f"{path}/{filename_with_extension}")
+    video_frames = Preprocessor.read_video_and_extract_face(f"{path}/{filename_with_extension}")
+
+    # Combine the frames
+    combined_frame = Preprocessor.combine_video_frames(video_frames)
+
+    # Save the combined image
     cv2.imwrite(combined_image_path, combined_frame)
     print(f"\t* Saved {combined_image_path}.")
